# Table performance improvements