--- a/MigraDocCore.DocumentObjectModel/MigraDoc.DocumentObjectModel.Visitors/MergedCellList.cs
+++ b/MigraDocCore.DocumentObjectModel/MigraDoc.DocumentObjectModel.Visitors/MergedCellList.cs
@@ -32,6 +32,7 @@
 
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using MigraDocCore.DocumentObjectModel.IO;
 using MigraDocCore.DocumentObjectModel.Tables;
 using MigraDocCore.DocumentObjectModel.Visitors;
@@ -42,49 +43,8 @@ namespace MigraDocCore.DocumentObjectModel.Visitors
   /// <summary>
   /// Represents a merged list of cells of a table.
   /// </summary>
-  public class MergedCellList : ArrayList
+  public class MergedCellList : List<Cell>
   {
-
-    /// <summary>
-    /// Enumerator that can iterate through the MergedCellList.
-    /// </summary>
-    public class Enumerator : IEnumerator
-    {
-      internal Enumerator(MergedCellList list)
-      {
-        this.list = list;
-      }
-
-      #region IEnumerator Members
-
-      public void Reset()
-      {
-        this.index = -1;
-      }
-
-      public Cell Current
-      {
-        get
-        {
-          return (Cell)this.list[this.index];
-        }
-      }
-
-      object IEnumerator.Current
-      {
-        get { return Current; }
-      }
-
-      public bool MoveNext()
-      {
-        return ++index < this.list.Count;
-      }
-      #endregion
-
-      MergedCellList list;
-      int index = -1;
-    }
-
     /// <summary>
     /// Enumeration of neighbor positions of cells in a table.
     /// </summary>
@@ -142,14 +102,6 @@ namespace MigraDocCore.DocumentObjectModel.Visitors
         }
       }
       return false;
-    }
-
-    /// <summary>
-    /// Gets the Enumerator for this list.
-    /// </summary>
-    public override IEnumerator GetEnumerator()
-    {
-      return new Enumerator(this);
     }
 
     /// <summary>