--- a/MigraDocCore.Rendering/MigraDoc.Rendering/TableRenderer.cs
+++ b/MigraDocCore.Rendering/MigraDoc.Rendering/TableRenderer.cs
@@ -569,22 +569,27 @@ namespace MigraDocCore.Rendering
       XUnit maxBottomBorderPosition = lastPos + minMergedFormattedCell.InnerHeight;
       maxBottomBorderPosition += CalcBottomBorderWidth(minMergedCell);
 
+      // Note: Caching the indices does speed up this function for large tables greatly.
+      var minMergedCellRowIndex = minMergedCell.Row.Index;
+      var minMergedCellMergeDown = minMergedCell.MergeDown;
+      var mergedIndexPlusDown = minMergedCellRowIndex + minMergedCellMergeDown;
       foreach (Cell cell in this.mergedCells)
       {
-        if (cell.Row.Index > minMergedCell.Row.Index + minMergedCell.MergeDown)
+        var rowIndex = cell.Row.Index;
+        if (rowIndex > mergedIndexPlusDown)
           break;
 
-        if (cell.Row.Index + cell.MergeDown == minMergedCell.Row.Index + minMergedCell.MergeDown)
+        if (rowIndex + cell.MergeDown == mergedIndexPlusDown)
         {
           FormattedCell formattedCell = (FormattedCell)this.formattedCells[cell];
-          XUnit topBorderPos = (XUnit)this.bottomBorderMap[cell.Row.Index];
+          XUnit topBorderPos = (XUnit)this.bottomBorderMap[rowIndex];
           XUnit bottomBorderPos = topBorderPos + formattedCell.InnerHeight;
           bottomBorderPos += CalcBottomBorderWidth(cell);
           if (bottomBorderPos > maxBottomBorderPosition)
             maxBottomBorderPosition = bottomBorderPos;
         }
       }
-      this.bottomBorderMap.Add(minMergedCell.Row.Index + minMergedCell.MergeDown + 1, maxBottomBorderPosition);
+      this.bottomBorderMap.Add(mergedIndexPlusDown + 1, maxBottomBorderPosition);
     }
 
     /// <summary>
@@ -614,20 +619,22 @@ namespace MigraDocCore.Rendering
       Cell minCell = null;
       foreach (Cell cell in this.mergedCells)
       {
-        if (cell.Row.Index == row)
+        var rowIndex = cell.Row.Index; // Note: Taking index only once speeds up large tables.
+        if (rowIndex <= row && rowIndex + cell.MergeDown >= row)
         {
-          if (cell.MergeDown == 0)
+          if (rowIndex == row && cell.MergeDown == 0)
           {
+            // Perfect match: non-merged cell in the desired row.
             minCell = cell;
             break;
           }
-          else if (cell.MergeDown < minMerge)
+          else if (rowIndex + cell.MergeDown - row < minMerge)
           {
-            minMerge = cell.MergeDown;
+            minMerge = rowIndex + cell.MergeDown - row;
             minCell = cell;
           }
         }
-        else if (cell.Row.Index > row)
+        else if (rowIndex > row)
           break;
       }
       return minCell;
@@ -644,11 +651,12 @@ namespace MigraDocCore.Rendering
       int lastConnectedRow = row;
       foreach (Cell cell in this.mergedCells)
       {
-        if (cell.Row.Index <= lastConnectedRow)
+        var index = cell.Row.Index; // Note: Caching index here for speedup for large tables.
+        if (index <= lastConnectedRow)
         {
           int downConnection = Math.Max(cell.Row.KeepWith, cell.MergeDown);
-          if (lastConnectedRow < cell.Row.Index + downConnection)
-            lastConnectedRow = cell.Row.Index + downConnection;
+          if (lastConnectedRow < index + downConnection)
+            lastConnectedRow = index + downConnection;
         }
       }
       return lastConnectedRow;