Made some performance improvement to rendering large tables. Some of these are from the 1.50 version, but I also added some low-hanging fruit I discovered with profiling. None of the changes should have an effect on anything other than performance.

For 2000 rows, the execution time is 8.8% of the original code; for 5000 rows, the execution time is 4.7% (local tests on my system). The speed improvements seem to grow the bigger the table gets. Yes, this means 3.5s vs. 39.4s for 2000 rows and 11.2s vs. 235s for 5000 rows, which is why we desperately need this.

Updating to .NET 5 actually made a large speed-up as well; it seems that the relative improvement of my updated version was even boosted more by that, because with the previous version, it was "only" 11%/6.3%. The .NET update improves the original code's execution time by about one third, but my code by half, so I'm quite happy about that.

NOTE: I have created a manual unit test for speed tests; if you want I can add it to the pull request, but I didn't because to make things easy, I just added a single test to the PdfSharpCore.Test assembly, but it's actually a Migradoc test.